+++ LogReader.Tests.MSTest/Usings.cs
@@ -1,2 +1,1 @@
 ﻿global using Microsoft.VisualStudio.TestTools.UnitTesting;
-global using Microsoft.VisualStudio.TestTools.UnitTesting.AppContainer;
