+++ LogReader.sln
@@ -10,47 +10,17 @@ EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
-		Debug|arm64 = Debug|arm64
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
-		Release|arm64 = Release|arm64
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Debug|arm64.ActiveCfg = Debug|arm64
-		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Debug|arm64.Build.0 = Debug|arm64
-		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Debug|x64.ActiveCfg = Debug|x64
-		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Debug|x64.Build.0 = Debug|x64
-		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Debug|x86.ActiveCfg = Debug|x86
-		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Debug|x86.Build.0 = Debug|x86
 		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Release|arm64.ActiveCfg = Release|arm64
-		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Release|arm64.Build.0 = Release|arm64
-		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Release|x64.ActiveCfg = Release|x64
-		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Release|x64.Build.0 = Release|x64
-		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Release|x86.ActiveCfg = Release|x86
-		{6D93FFC3-3CE5-43B9-AEB1-97E1F732F7BA}.Release|x86.Build.0 = Release|x86
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Debug|Any CPU.Build.0 = Debug|x64
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Debug|arm64.ActiveCfg = Debug|arm64
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Debug|arm64.Build.0 = Debug|arm64
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Debug|x64.ActiveCfg = Debug|x64
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Debug|x64.Build.0 = Debug|x64
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Debug|x86.ActiveCfg = Debug|x86
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Debug|x86.Build.0 = Debug|x86
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Release|Any CPU.ActiveCfg = Release|x64
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Release|Any CPU.Build.0 = Release|x64
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Release|arm64.ActiveCfg = Release|arm64
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Release|arm64.Build.0 = Release|arm64
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Release|x64.ActiveCfg = Release|x64
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Release|x64.Build.0 = Release|x64
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Release|x86.ActiveCfg = Release|x86
-		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Release|x86.Build.0 = Release|x86
+		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D03DF3F1-202B-4BB3-8C2D-A19624EDEF2A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
